remove unecessary hierarchy
