+++ tchannel/handler.py
@@ -19,73 +19,6 @@
 # THE SOFTWARE.
 from __future__ import absolute_import
 
-from .errors import InvalidChecksumError
-from .errors import StreamingError
-from .messages import ErrorCode
-from .messages import Types
-
-
-class BaseRequestHandler(object):
-    """A minimal RequestHandler skeleton.
-
-    This implements a minimal base RequestHandler that implements methods
-    which should be mostly the same between implementations. Implementations
-    must inherit this class and override at least ``handle_call``.
-    """
-
-    _HANDLER_NAMES = {
-        Types.PING_REQ: 'ping',
-        Types.CALL_REQ: 'pre_call',
-        Types.CALL_REQ_CONTINUE: 'pre_call'
-    }
-
-    def handle(self, message, connection):
-        # TODO assert that the handshake was already completed
-        assert message, "message must not be None"
-
-        if message.message_type not in self._HANDLER_NAMES:
-            # TODO handle this more gracefully
-            raise NotImplementedError("Unexpected message: %s" % str(message))
-
-        handler_name = "handle_" + self._HANDLER_NAMES[message.message_type]
-        return getattr(self, handler_name)(message, connection)
-
-    def handle_pre_call(self, message, connection):
-        """Handle incoming request message including CallRequestMessage and
-        CallRequestContinueMessage
-
-        This method will build the User friendly request object based on the
-        incoming messages.
-
-        It passes all the messages into the message_factory to build the init
-        request object. Only when it get a CallRequestMessage and a completed
-        arg_1=argstream[0], the message_factory will return a request object.
-        Then it will trigger the async call_handle call.
-
-        :param message: CallRequestMessage or CallRequestContinueMessage
-        :param connection: tornado connection
-        """
-        try:
-            req = connection.request_message_factory.build(message)
-            # message_factory will create Request only when it receives
-            # CallRequestMessage. It will return None, if it receives
-            # CallRequestContinueMessage.
-            if req:
-                self.handle_call(req, connection)
-
-        except (InvalidChecksumError, StreamingError) as e:
-            connection.send_error(
-                ErrorCode.bad_request,
-                e.message,
-                message.id,
-            )
-        except Exception:
-            connection.send_error(
-                ErrorCode.unexpected,
-                "An unexpected error has occurred.",
-                message.id,
-            )
-
 
 class CallableRequestHandler(object):
     """An adapter from a function to a RequestHandler."""
